Change double type to int in var and methods

+++ ChocoShop/src/main/java/Basket.java
@@ -19,11 +19,11 @@ public class Basket {
             return quantity;
         }
 
-        public double getFullPrice() {
+        public int getFullPrice() {
             return fullPrice;
         }
 
-        private double fullPrice;
+        private int fullPrice;
 
         public OrderLine(Chocolate chocolate, int quantity) {
             this.chocolate = chocolate;
