+++ ChocoShop/src/main/java/OrderCalculator.java
@@ -1,6 +1,6 @@
 public class OrderCalculator {
-    public static double calculateOrderPrice(Basket basket){
-        double finalCost = 0;
+    public static int calculateOrderPrice(Basket basket){
+        int finalCost = 0;
         for (Basket.OrderLine orderLine : basket.getOrder()){
             finalCost+=TaxesCalculator.calculateTaxes(orderLine.getFullPrice());
         }
